requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
pydantic-settings>=2.1.0

# External dependencies
openai>=1.12.0
cachetools>=5.3.3
redis>=5.0.1
prometheus-client>=0.19.0

# Testing dependencies
pytest>=8.3.5
pytest-cov>=4.1.0
pytest-asyncio>=0.23.5
python-dotenv>=1.0.0
httpx>=0.28.1
ruff>=0.3.0